refactor: moved mcsd resources enum and simplified get_resources function
