+++ tests/services/update/test_update_client_service.py
@@ -1,10 +1,10 @@
 from typing import Sequence
 from unittest.mock import ANY, MagicMock
 from app.config import ConfigExternalCache
+from app.models.fhir.types import McsdResources
 from app.services.update.cache.provider import CacheProvider
 from app.services.update.update_client_service import (
     UpdateClientService,
-    McsdResources,
 )
 from fhir.resources.R4B.bundle import Bundle, BundleEntry, BundleEntryRequest
 
@@ -76,10 +76,10 @@ def test_cleanup() -> None:
         assert called_bundle.entry is not None
         assert len(called_bundle.entry) == 2
         assert isinstance(called_bundle.entry[0], BundleEntry)
-        assert isinstance(called_bundle.entry[0].request, BundleEntryRequest) # type: ignore[attr-defined]
-        assert called_bundle.entry[0].request.method == "DELETE" # type: ignore[attr-defined]
-        assert called_bundle.entry[0].request.url is not None # type: ignore[attr-defined]
-        split = called_bundle.entry[0].request.url.split("/") # type: ignore[attr-defined]
+        assert isinstance(called_bundle.entry[0].request, BundleEntryRequest)  # type: ignore[attr-defined]
+        assert called_bundle.entry[0].request.method == "DELETE"  # type: ignore[attr-defined]
+        assert called_bundle.entry[0].request.url is not None  # type: ignore[attr-defined]
+        split = called_bundle.entry[0].request.url.split("/")  # type: ignore[attr-defined]
         assert split[0] == resource.value
         assert (
             split[1] == "resource_1?_cascade=delete"
